requirements.txt
requests
python-dotenv
google-generativeai
openpyxl
gophish
urllib3
six
matplotlib